try some different stuff with the comms to fix the failing tests

+++ src/helics/shared_api_library/MessageFiltersExport.cpp
@@ -275,7 +275,7 @@ helics_filter helicsFederateGetFilterByIndex (helics_federate fed, int index, he
 		if (!filt->filtptr)
 		{
             auto fobj = reinterpret_cast<helics::FedObject *> (fed);
-			if (index < fobj->filters.size())
+			if (index < static_cast<int>(fobj->filters.size()))
 			{
                 filt->filtptr = fobj->filters[index]->filtptr;
 			}
